Completes ex 8.3, adds author schema & all query
This commit implements an allAuthors query which returns an array of
all authors in the collection. It also modifies the author schema
added in a previous commit to include a 'bookCount' field, which is
the total number of books by that author in the books collection. A
resolver for the Author schema is added to resolve the bookCount
field, as it is not part of the data source for Authors.