Completes ex. 8.7, adds editAuthor mutation
This commit implements a mutation which changes the birth year of an
author in the collection. This mutation, titled editAuthor, takes two
parameters: the name of the author to update and the new birth year.
As birth year is not required for the Author schema, it is also not
a required parameter for the mutation - if not passed, the author's
birth year will be set to null. If the name is not found in the
author collection, no changes are made to the author collection and
the mutation returns null.