Completes ex. 8.10, implements add book form
This commit implements the add book form. It adds a CREATE_BOOK
mutation to queries.js, using the data entered in the books form. The
published entry is set to type 'number', to ensure the data entered
is a number - in the query, the string data is converted to an INT
via parseInt.

A useEffect hook is used to reset the fields of the form, using the
result parameter of the useMutation hook's return. The useEffect runs
when result.loading or result.error updates, and if both values are
falsey it resets all of the input fields.

The useMutation call in the NewBook component includes parameters to
update the cached values for the authors view and the books view when
run. It also includes error handling, with the error message
displayed in window.alert.