Completes ex. 8.11, adds update for birth year
This commit adds a new Birthyear component which is rendered by the
Authors component. This component includes a form for updating an
author's birth year, using the editAuthor mutation. The mutation is
defined in queries.js. If the author entered in the form is not
present in the collection, a window.alert notifies the user of this.
The form is only reset after a successful update. The authors cache
is also updated after calling the mutation.