Completes ex. 8.13, connects library to DB
This commit connects the backend of the library application to a
mongoDB database. It adds mongo schema for books and authors, and
modifies the queries and mutations to save to and retrieve from the
database. This commit also modifies the GraphQL Book schema author
member to be an Author object, rather than a String.

Per the guidelines of the exercise, the following are not yet
implemented:
* allBooks query with parameters
* bookCount field of author object
* author field of book object
* editAuthor mutation