Completes ex. 8.16, adds user management
This commit adds user management and authentication to the library
app. Users are created with a username and required favorite genre.
There is a login mutation which returns a token if a valid username
and the correct password are entered (per the exercise instructions,
a common password is hardcoded). A new 'me' quey returns the logged
in user. The addBook and editAuthor now require a valid token or will
throw an error.