Completes ex 8.18, adds login to library frontend
This commit adds a login form to the library frontend, routed to
/login. The login alerts on an incorrect login. If the login
succeeds, the token is saved to react context and to local storage.
The token is used from local storage to add to the apollo client
request headers configured in main.jsx. When the react context does
not show a user logged in (context is null or falsey), the button for
add book is not displayed, nor is the form to change an author's
birth year on the authors page. When the context contains a token,
the button to login is not diplayed, and a logout button is instead
displayed.

This commit also fixes an issue with the CREATE_BOOK query to reflect
changes made to the backend in previous exercises.