Completes ex. 8.19, adds genre filter to books
This commit adds a set of buttons to the Books component, one for
each genre present in the books list as well as an 'all genres'
button. Clicking a button filters the books list to only show books
of that genre. The list of buttons is generated using the SET
object.