Completes ex. 8.20, adds Recommendations view
This commit adds a new Recommendations component, which displays the
books which have as a genre the logged-in user's favorite genre.
Implementing this view requires modifications to the NavMenu
component, routing, and the addition of a FIND_USER client query that
implements the me server query.

This commit also adds a BooksList component, which takes as a prop an
array of books and renders the list. This component is rendered in
both the Books component (which now only queries the server for books
and handles the filtering by genre) and the Recommendations component
(acting similarly to the Books component, except filtering by the
user's favorite genre rather than implementing the option to choose
genre to filter with).