Completes ex 8.21, fetches books by genre from DB
This commit updates the functionality of the genre filters on the
book list so that instead of filtering in React it sends a genre
parameter to the server in the query and receives back a set of books
filtered based on the parameter.

This commit also fixes a lot of issues with apollo's cache present in
previous commits.