Completes ex 8.22, makes books view stay updated
This commit sets the fetch policy for the queries that populate the
BooksList component to no-cache, so that every time the BooksList
component is re-rendered (e.g., by pressing a genre filter button)
the view is kept up to date. This includes both the list of books and
the list of genres.

This commit also removes several now-unnecessary refetchQuery members
from the useMutation used in the newBook component.