Completes ex 8.26, resolves n+1 for bookCount
This commit resolves an n+1 issue in the Author schema bookCount
parameter. It accomplishes this by adding bookCount as an Int member
of the Author schema. The bookCount member is incremented by 1
whenever a new book is added, so long as the Author member has a
non-null bookCount. For new Authors, the addBook mutation initializes
the value to 1. In addition, the validity of the bookCount is
maintained in two ways. First, if the bookCount is null when it is
queried, it will query the Books collection to find the correct count
and save that value to the database (this can cause a single n+1
query when bookCount is run for all authors if none have a populated
bookCount). Second, an AuditAuthorBookCount mutation is added, which
takes an Author's name as a variable and so long as the Author is in
the collection it searches the Books collection to find the correct
count and updates the Author's bookCount value in the database.